Move Key_File create functions to libeconf.c
Since those functions are part of the user API they should be
defined in libeconf.c
