+++ include/keyfile.h
@@ -33,8 +33,6 @@ typedef struct Key_File {
   char delimiter, comment;
 } Key_File;
 
-Key_File newKeyFile(char delimiter, char comment);
-Key_File newIniFile();
 void key_file_append(Key_File *key_file);
 
 /* GETTERS */
