+++ src/keyfile.c
@@ -31,29 +31,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-// Create a new Key_File. Allocation is based on KEY_FILE_DEFAULT_LENGTH
-// defined in include/defines.h
-Key_File newKeyFile(char delimiter, char comment) {
-  Key_File key_file;
-
-  key_file.alloc_length = KEY_FILE_DEFAULT_LENGTH;
-  key_file.length = 0;
-  key_file.delimiter = delimiter;
-  key_file.comment = comment;
-
-  key_file.file_entry = malloc(KEY_FILE_DEFAULT_LENGTH * sizeof(struct file_entry));
-
-  for (size_t i = 0; i < KEY_FILE_DEFAULT_LENGTH; i++) {
-    set_default_value(&key_file, i);
-  }
-
-  return key_file;
-}
-
-Key_File newIniFile() {
-  return newKeyFile('=', '#');
-}
-
 void key_file_append(Key_File *kf) {
   if(kf->length++ >= kf->alloc_length) {
     kf->alloc_length++;
